Refactor selected databases to store its data in settings
