+++ app/models/solr_document.rb
@@ -63,6 +63,10 @@ class SolrDocument
     document.key?(:modsxml)
   end
 
+  use_extension(SelectedDatabase) do |document|
+    Settings.selected_databases[document.id].present?
+  end
+
   sw_field_semantics = {
     title: %w[title_display eds_title],
     author: 'author_display',
