+++ config/settings.yml
@@ -84,3 +84,47 @@ LIB_GUIDES:
   MINI_BENTO_ENABLED: false
   NUM_RESULTS: 3
   SITE_ID: 12345
+
+selected_databases:
+   "5749286":
+     subjects: ["General", "Multidisciplinary"]
+     description: "Multidisciplinary database of over 4,650 mostly scholarly publications, many full-text; a great place to start your research."
+   "6494821":
+     subjects: ["Art"]
+     description: "Bibliographic database on art and related disciplines; also indexes art reproductions. Coverage begins in 1984, but full-text starts in 1997; a related database, Art Retrospective, indexes articles from 1929-1984"
+     see_also: { id: "6666306", text: "Art Retrospective" }
+   "6631086":
+     subjects: ["Business", "Economics"]
+     description: "Provides full-text for over 10,000 scholarly business journals and other sources for business and economics related topics."
+   "3964904":
+     subjects: ["Education"]
+     description: "This database provides users with ready access to an extensive body of education-related literature."
+   "8666689":
+     subjects: ["World history"]
+     description: "Covers the history of the world (excluding the U.S. and Canada) from 1450 to the present, 1955-; some full text."
+     see_also: { id: "8666684", text: "America: History and Life" }
+   "6758881":
+     subjects: ["Sciences", "Humanities", "Social Sciences", "Citation Indexes"]
+     description: "Web of Knowledge provides access to information on science, social sciences, and arts and humanities, as well as search and analysis (citation) tools."
+   "3942381":
+     subjects: ["Multidisciplinary", "Newspaper articles"]
+     description: "Provides full-text documents from over 5,900 news, business, legal, medical, and reference publications."
+     see_also: { id: "9688486", text: "ProQuest News & Newspapers" }
+   "497192":
+     subjects: ["Literature", "Humanities"]
+     description: "A bibliography of journal articles, books and dissertations covering literature, language and linguistics, film, folklore, literary theory & criticism, dramatic arts, as well as the historical aspects of printing and publishing."
+   "8785205":
+     subjects: ["General", "Multidisciplinary"]
+     description: "From business and political science to literature and psychology, access to a wide range of popular academic subjects; includes more than 5,000 titles (over 3,500 in full text) from 1971 forward."
+   "4291719":
+     subjects: ["Psychology"]
+     description: "Contains citations and summaries of journal articles, book chapters, books, and technical reports in the field of psychology and psychological aspects of many related disciplines, 1840-present."
+   "375121":
+     subjects: ["Music"]
+     description: "Comprehensive guide to publications on music from all over the world, with abstracts written in English, 1967-present."
+   "8545966":
+     subjects: ["Science", "Social Sciences"]
+     description: "This large abstract and citation database from Elsevier includes international sources in the scientific, technical, medical and social sciences fields and, more recently, the arts and humanities."
+   "488729":
+     subjects: ["Sociology", "Social Sciences"]
+     description: "Access to the international literature in sociology and related disciplines in the social and behavioral sciences, 1953-present."
